Move tests to a sister directory of `src`
- This fits better with how tests should be used
- Include a pyproject.toml for pytest usage
