+++ .github/workflows/unit_tests.yml
@@ -29,15 +29,6 @@ jobs:
           extra-specs: |
             python=${{ matrix.python-version }}
 
-      - name: Lint with flake8
-        run: |
-          # Initialize micromamba and activate como_tests environment
-          eval "$(micromamba shell hook --shell=bash)" && micromamba activate como_tests
-          # stop the build if there are Python syntax errors or undefined names
-          flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
-          # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-          flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
-
       - name: Run tests
         run: |
           # Initialize micromamba and activate como_tests environment
