+++ main/tests/environment.yaml
@@ -14,7 +14,6 @@ dependencies:
   - pytest-localftpserver==1.1.3
   - pytest-mock==3.8.2
   - pytest==7.1.3
-  - pytest==7.1.3
   - rpy2==3.4.5
   - tqdm==4.64.1
   - pip==22.1.2
