+++ main/tests/inputs/test_proteomics_input.csv
