+++ main/tests/test_execution.py
@@ -1,0 +1,3 @@
+def test_imports():
+    from src.utilities import Compartments, stringlist_to_list, split_gene_expression_data
+    from src.project import configs
