+++ main/tests/test_model_creation.py
@@ -1,0 +1,3 @@
+"""
+The purpose of this file is to test the result of creating a constraint-based metabolic model with the provided naiveB cell data
+"""
