+++ pyproject.toml
@@ -1,0 +1,4 @@
+[tool.pytest.ini_options]
+pythonpath = [
+    "main/src"
+]
